+++ Clippit.Cli/CliJsonContext.cs
@@ -1,5 +1,6 @@
 using System.Text.Json.Serialization;
 using Clippit.Cli.Commands.Common.Verify;
+using Clippit.Cli.Commands.Excel.Create;
 using Clippit.Cli.Commands.Pptx.Build;
 using Clippit.Cli.Commands.Pptx.Split;
 using Clippit.Cli.Commands.Pptx.Verify;
@@ -40,6 +41,8 @@ namespace Clippit.Cli;
 [JsonSerializable(typeof(WordBuildResult))]
 [JsonSerializable(typeof(WordBuildEntryResult))]
 [JsonSerializable(typeof(WordBuildInitResult))]
+[JsonSerializable(typeof(WorkbookDefinition))]
+[JsonSerializable(typeof(CreateResult))]
 [JsonSerializable(typeof(ErrorResult))]
 [JsonSourceGenerationOptions(
     WriteIndented = false,